huggingface-dashboard/requirements.txt
bayesian_optimization==1.4.2
great_expectations==0.14.12
hopsworks==3.0.5
hsml==3.0.3
joblib==1.2.0
matplotlib==3.5.3
modal==0.4.1
modal_client==0.44.679
numpy==1.21.6
pandas==1.3.5
praw==7.6.1
pytest==7.2.0
python-dotenv==0.21.0
scikit_learn==1.2.0
scipy==1.7.3
seaborn==0.12.2
sentence_transformers==2.2.2
shap==0.41.0
streamlit==1.16.0
tqdm==4.64.1
transformers==4.25.1
xgboost==1.6.2
